changed the texturebase into an interface. changed some values from int and ushort to byte
I don't think anything will have 255 cults/palettes.  or even 127 as i think somehting is using an sbyte
cards probably have the most cluts that i know of. as it has one for each card and the card back.

added alternative color struct for 16 bit color. that will let you read it with out converting to 32 bit color. Though most of the code is expecting a 32 bit color struct. if I could make my own struct I'd have both using the same interface. Though I might do that anyway. and just have the custom struct pass implict cast to xna's color.

+++ Core/Font.cs
@@ -85,7 +85,7 @@ namespace OpenVIII
             var bufferTDW = aw.GetBinaryFile("sysfnt.tdw");
             var tim = new TDW(bufferTDW);
             charWidths = tim.CharWidths;
-            menuFont = tim.GetTexture((ushort)ColorID.White);
+            menuFont = tim.GetTexture((byte)ColorID.White);
             
         }
